requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
jinja2==3.1.6
google-cloud-aiplatform
google-cloud-texttospeech
aiohttp
websockets
python-dotenv==1.0.1